app/requirements.txt
numpy==1.19.5
imutils==0.5.4
pandas==1.4.1
Pillow==8.4.0
opencv-python==4.6.0.66
scikit-learn==1.0.2
TensorFlow==2.4.1
keras==2.4.0
keras-facenet==0.3.2